fix resource_url caused bugs

+++ app/controllers/application_controller.rb
@@ -18,13 +18,13 @@ class ApplicationController < ActionController::Base
     random_branch_question = nil
     question_pack.questions.each do |question|
       if question.branch_questions.any?
-        random_branch_question = question.branch_questions[0]  #找到某个小题的resource_url
+        random_branch_question = question.branch_questions.where("resource_url is not null")[0]  #找到某个小题的resource_url
         break
       end
     end
 
     #删除作业文件夹
-    if random_branch_question.present?
+    if random_branch_question.present? && random_branch_question.resource_url.present?
       branch_question_resource_path = "#{Rails.root}/public/" + random_branch_question.resource_url
       question_pack_dir = File.dirname(branch_question_resource_path)
       FileUtils.remove_dir question_pack_dir if Dir.exist? question_pack_dir
