+++ app/controllers/questions_controller.rb
@@ -36,8 +36,10 @@ class QuestionsController < ApplicationController
     Question.transaction do
       if question.destroy
         branch_questions.each do |bq|
-          resource_path = (Rails.root.to_s + "/public" + bq.resource_url)
-          File.delete resource_path if File.exists?(resource_path)
+          if bq.resource_url.present?
+            resource_path = (Rails.root.to_s + "/public" + bq.resource_url)
+            File.delete resource_path if File.exists?(resource_path)
+          end
         end
       end
       flash[:notice]="删除成功"
@@ -56,7 +58,7 @@ class QuestionsController < ApplicationController
           share_question = ShareQuestion.create({:user_id => current_user.id, :name => question.name, :types => question.types, :cell_id => question.cell_id, :episode_id => question.episode_id})
           if share_question
             question.branch_questions.each do |bq|
-              new_resource_url = copy_file(SAHRE_MEDIA_PATH, question_pack, bq, bq.resource_url) #分享的时候，拷贝音频
+              new_resource_url = copy_file(SAHRE_MEDIA_PATH, question_pack, bq, bq.resource_url) if bq.resource_url.present? #分享的时候，拷贝音频
               share_question.share_branch_questions.create({:content => bq.content, :resource_url => new_resource_url})
             end
           end
@@ -94,7 +96,7 @@ class QuestionsController < ApplicationController
     Question.transaction do
       share_branch_questions.each do |sbq|
         branch_question = @question.branch_questions.create({:content => sbq.content})
-        new_resource_url = copy_file(MEDIA_PATH, @question_pack, branch_question, sbq.resource_url) #引用的时候，拷贝音频
+        new_resource_url = copy_file(MEDIA_PATH, @question_pack, branch_question, sbq.resource_url) if sbq.resource_url.present? #引用的时候，拷贝音频
         branch_question.update_attribute(:resource_url, new_resource_url)
         if branch_question == @question.branch_questions.first
           @question.update_attribute(:name, branch_question.content.length > 38 ? branch_question.content[0..35] + "..." : branch_question.content)
@@ -105,7 +107,7 @@ class QuestionsController < ApplicationController
 
   private
 
-  #引用的时候，拷贝音频
+  #分享或者引用的时候，拷贝音频
   def copy_file(media_path, question_pack, branch_question, source_resource_url)
     full_media_path = "/public" + media_path % question_pack.id
     question_pack_folder = Rails.root.to_s + full_media_path
